Adds more reference to interpretation.md
When I was learning this library, the documentation for the interpreter did not jump out to me immediately. However, I find it to be a vital tool in my applications. I think issue #264 also points to this difficulty. I'm proposing this change because I think it would make the learning process quicker and more smoother.